ensure that relative paths in asv.conf resolve
Apparently asv uses directories relative to the location of the config file, so in order
for https://github.com/MDAnalysis/benchmarks/wiki/benchmark-on-c3potato to still work
we need a asv.conf.json at the top level of the benchmark directory.

cron.sh was adjusted so that we cd in the correct directories so that relative
paths resolve properly.